fix(sasjs-add): fix tests and restructure code
