+++ src/types/TargetScope.ts
@@ -1,0 +1,4 @@
+export enum TargetScope {
+  Global = 'Global',
+  Local = 'Local'
+}
